# Add support for compressed assemblies in APK