--- a/build-tools/xaprepare/xaprepare/xaprepare.csproj
+++ b/build-tools/xaprepare/xaprepare/xaprepare.csproj
@@ -157,6 +157,8 @@
     <Compile Include="ThirdPartyNotices\aapt2.cs" />
     <Compile Include="ThirdPartyNotices\bundletool.cs" />
     <Compile Include="ThirdPartyNotices\Java.Interop.cs" />
+    <Compile Include="ThirdPartyNotices\K4os.Compression.LZ4.cs" />
+    <Compile Include="ThirdPartyNotices\lz4.cs" />
     <Compile Include="ThirdPartyNotices\mono.cs" />
     <Compile Include="ThirdPartyNotices\opentk.cs" />
     <Compile Include="ThirdPartyNotices\proguard.cs" />