--- a/src/Xamarin.Android.Build.Tasks/Utilities/CompressedAssemblyInfo.cs
+++ b/src/Xamarin.Android.Build.Tasks/Utilities/CompressedAssemblyInfo.cs
@@ -1,0 +1,16 @@
+namespace Xamarin.Android.Tasks
+{
+	class CompressedAssemblyInfo
+	{
+		public const string CompressedAssembliesInfoKey = "__CompressedAssembliesInfo";
+
+		public uint FileSize { get; }
+		public uint DescriptorIndex { get; set; }
+
+		public CompressedAssemblyInfo (uint fileSize)
+		{
+			FileSize = fileSize;
+			DescriptorIndex = 0;
+		}
+	}
+}