… Documentation/release-notes/4686.md
… src/Xamarin.Android.Build.Tasks/Tests/Xamarin.Android.Build.Tests/IncrementalBuildTest.cs
… src/Xamarin.Android.Build.Tasks/Xamarin.Android.Common.targets
… src/monodroid/CMakeLists.txt
… src/monodroid/jni/embedded-assemblies.cc
… tests/apk-sizes-reference/Xamarin.Forms_Performance_Integration-Signed-Release-Bundle.apkdesc