@jonpryor 2020-05-18
@JonDouglas chimed in today and believes that we should have assembly compression enabled by default.