@brendanzagaeski 2020-05-20
*Draft release notes*

During the finalization of this PR before merge, when you get a chance, you can add a release note for it to `Documentation/release-notes/4686.md` as part of the PR. Here's a rough draft that has some `??` items to indicate missing info. The other info might also not be 100% accurate, so feel free to edit liberally. Thanks!

````
### Smaller app package sizes

This version introduces compression of managed assemblies by default for Release
configuration builds, resulting in significantly smaller APK and App Bundle
sizes.  Assemblies are compressed with the [LZ4][lz4] algorithm during builds
and then decompressed on device during app startup.

For a small example Xamarin.Forms application, this reduced the APK size from
about 23 megabytes to about 17 megabytes while only increasing the time to
display the first page of the app from about 780 milliseconds to about 790
milliseconds in the least favorable configuration.

If needed, the new behavior can been disabled for a particular project by
setting the `AndroidEnableAssemblyCompression` MSBuild property to `false` in
the _.csproj_ file:

```xml
<PropertyGroup>
  <AndroidEnableAssemblyCompression>false</AndroidEnableAssemblyCompression>
</PropertyGroup>
```

> [!NOTE]
> This feature is intended to replace the older **Bundle assemblies into native
> code** Visual Studio Enterprise feature for purposes of app size savings.  The
> `AndroidEnableAssemblyCompression` property takes precedence if both features
> are enabled.  Project authors who no longer need the **Bundle assemblies into
> native code** feature enabled can now disable it or remove the
> `BundleAssemblies` MSBuild property from the _.csproj_ file:
>
> ```diff
>  <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Release|AnyCPU' ">
>    <DebugSymbols>True</DebugSymbols>
>    <DebugType>portable</DebugType>
>    <Optimize>True</Optimize>
>    <OutputPath>bin\Release\</OutputPath>
>    <DefineConstants>TRACE</DefineConstants>
>    <ErrorReport>prompt</ErrorReport>
>    <WarningLevel>4</WarningLevel>
>    <AndroidManagedSymbols>true</AndroidManagedSymbols>
>    <AndroidUseSharedRuntime>False</AndroidUseSharedRuntime>
>    <AndroidLinkMode>SdkOnly</AndroidLinkMode>
>    <EmbedAssembliesIntoApk>True</EmbedAssembliesIntoApk>
> -  <BundleAssemblies>true</BundleAssemblies>
>  </PropertyGroup>
> ```

#### Background information

For comparison, for the small test Xamarin.Forms application, the **Bundle
assemblies into native code** feature reduces the APK size from about 23
megabytes to about ?? megabytes while ??increasing?? the time to display the
first page of the app from about 780 milliseconds to about ?? milliseconds in
the least favorable configuration.

[lz4]: https://github.com/lz4/lz4
````