@dellis1972 2020-05-21
The failing test does not seem to be an issue caused directly by this PR. 
Further investigation reviled an existing issue with incremental builds when using `Hybrid` AOT. 
Because `Hybrid` AOT calls the `CilStrip` task during the `_BuildApkEmbed` target it updates the assembly. The next build after than WILL cause `_GenerateJavaStubs` to run because the `dlls` are newer than the `stamp` files. 

While `_BuildApkEmbed` doesn't run in that scenario (and does with this new code). I believe we should try to alter the code so that `CilStrip` does NOT cause the next build to do work it doesn't not need to. So that should be fixed in a separate PR.