@jonpryor 2020-05-22
I'm not quite sure I understand the use/interaction of `@(AndroidCustomMetaDataForReferences)` item group.  I had *thought* that the `%(AndroidSkipCompression)` metadata would be provided via e.g. `%(Reference.AndroidSkipCompression)` or `%(ProjectReference.AndroidSkipCompression)`.  Can those also be used?  Or is `@(AndroidCustomMetaDataForReferences)` needed?