@grendello 2020-05-22
> I'm not quite sure I understand the use/interaction of `@(AndroidCustomMetaDataForReferences)` item group. I had _thought_ that the `%(AndroidSkipCompression)` metadata would be provided via e.g. `%(Reference.AndroidSkipCompression)` or `%(ProjectReference.AndroidSkipCompression)`. Can those also be used? Or is `@(AndroidCustomMetaDataForReferences)` needed?

Either can be used. The item group simply makes it possible for the assemblies you don't control - coming from nugets for instance - as you wouldn't be able to add metadata items to them in any other way.