@jonpryor 2020-05-22
> Single ABI enabled

I find it surprising that `$(BundleAssemblies)` results in a *smaller* `.apk` size when only a single ABI is included in the `.apk`.  Not only that, but that `$(BundleAssemblies` is 11% smaller!

Any idea why this is the case?