@grendello 2020-05-22
> > Single ABI enabled
> 
> I find it surprising that `$(BundleAssemblies)` results in a _smaller_ `.apk` size when only a single ABI is included in the `.apk`. Not only that, but that `$(BundleAssemblies` is 11% smaller!
> 
> Any idea why this is the case?

`gzip` compression ratio is better than `lz4` but much slower - at level `-9` the former is able to process around 15 MB/s while the latter 46 MB/s. On the same machine, `gzip` decompresses at around 235 MB/s while `lz4` at 550 MB/s